requirements.txt
django-bootstrap-form==3.0
django-countries==1.5
django-debug-toolbar==0.11.0
django-filter==0.7
django-mptt-admin==0.1.6
django-mptt==0.6.0
django-reset==0.2.0
Django-Select2==4.2.1
django-suit==0.2.5
Django==1.5.4
djangorestframework==2.3.8
gunicorn==18.0
ipython==1.1.0
Pillow==2.2.1
pip-tools==0.3.4
psycopg2==2.5.1
python-memcached==1.53
pytz==2013.7
raven==3.5.0
six==1.4.1
sorl-thumbnail==11.12
South==0.8.2
sqlparse==0.1.10
